server/requirements.txt
fastapi==0.115.14
uvicorn==0.25.0
pymongo==4.10.1
python-multipart==0.0.20
python-dotenv==1.0.0
